+++ src/Amani.ImportadosERP.Application/Interfaces/IDashboardOperacionalRepository.cs
@@ -3,7 +3,7 @@ namespace Amani.ImportadosERP.Application.Interfaces;
 public interface IDashboardOperacionalRepository
 {
     Task<int> ObterProdutosCadastradosAsync(DateTime dataReferencia);
-    Task<int> ObterEstoqueDisponivelTotalAsync(DateTime dataReferencia);
+    Task<decimal> ObterEstoqueDisponivelTotalAsync(DateTime dataReferencia);
     Task<(int Quantidade, decimal Valor)> ObterMercadoriasEmTransitoAsync(DateTime dataReferencia);
     Task<int> ObterComprasEmAbertoAsync(DateTime dataReferencia);
     Task<int> ObterProdutosPendentesRecebimentoAsync(DateTime dataReferencia);
